+++ release.clj
@@ -28,7 +28,8 @@
 (defn- git-push [params]
   (println "\nSyncing with github...")
   (when-not (and (zero? (:exit (git ["push" "origin" tag] {})))
-                 (zero? (:exit (git ["push" "origin"] {}))))
+                 (zero? (:exit (git ["push" "origin"] {})))
+                 (zero? (:exit (git ["push" "origin" "gh-pages"] {}))))
     (die 15 "\nCouldn't sync with github."))
   params)
 
@@ -65,7 +66,30 @@
 
 (defn- build-template "Create the standalone_template.html file" [params]
   (println "\nBuilding standalone_template.html...")
-  (b/process {:command-args ["clojure" "-X:build-template"]})
+  (when-not (zero? (:exit (b/process {:command-args ["clojure" "-A:cljs" "-X:build-template"]})))
+    (die 16 "\nCould not build standalone_template.html."))
+  params)
+
+(defn- build-standalone-examples [params]
+  (println "\nBuilding standalone HTML for GitHub Pages...")
+  (when-not (-> {:command-args ["clojure" "-X:cli" "standalone"
+                                (str "'" {:sources     [{:file/name "resources/mbrainz-schema/schema.edn"}
+                                                        {:file/name "resources/mbrainz-schema/enums.edn"}
+                                                        {:file/name "resources/mbrainz-schema/supplemental.edn"}]
+                                          :output-path "_site/mbrainz-schema.html"} "'")]}
+                b/process
+                :exit
+                zero?)
+    (die 17 "\nCould not create mbrainz-schema.html"))
+  (when-not (-> {:command-args ["clojure" "-X:cli" "standalone"
+                                (str "'" {:sources     [{:file/name "resources/schema-voyager-schema/schema.edn"}]
+                                          :output-path "_site/schema-voyager-schema.html"} "'")]}
+                b/process
+                :exit
+                zero?)
+    (die 18 "\nCould not create schema-voyager-schema.html"))
+  (when-not (zero? (:exit (git ["commit" "--all" "-m" "'Deploy updates'"] {:dir "./_site"})))
+    (die 19 "\nCould not commit GitHub Pages"))
   params)
 
 (defn- post-release-message "Suggest updating the docs to reference the new release." [params]
@@ -104,6 +128,7 @@
   * Ensure the tag is available on Github"
   [params]
   (check-release params)
+  (build-standalone-examples params)
   (git-push params)
   (post-release-message params)
   params)